fix: support multi-arch
Adapt install_plugin.sh to download the correct binary. Otherwise the
binary for amd64 systems will be downloaded on systems with a different
target architecture and the execution of helm cm-push throws an exec
format error.

If the release can not be found for the target architecture, the install
script throws an error.